Refactoring of imports
