+++ brevitas/quant_tensor.py
@@ -42,8 +42,7 @@ from collections import namedtuple
 
 import torch
 
-from brevitas.function.ops import ceil_ste
-from brevitas.function.ops_ste import round_ste
+from brevitas.function.ops_ste import round_ste, ceil_ste
 from brevitas.function.ops import max_uint
